Deals
Se agrega funcionalidad para la creación del deal.
